improve .gitconfig generation
Now the git and github params are stored in a YAML cache file
between runs, so .gitconfig is kept up to date without the user
needing to enter this info by hand each time they want to update
the .gitconfig file from its template.

The cache is kept in ~/.gitconfig.cache.