clean up and remove neovim from deploy
neovim is fine and all but not that useful for me now,
on servers it's always an old version, so the advantage
of plugins and tools that need 0.8+ or 0.9 is obliterated.
maybe will become useful in the future. it's only nice on
mac, but i already have vscode on mac.